feat(payment): 결제 UseCase 정의 및 금액 데이터 BigDecimal로 변경

+++ module-app/src/test/kotlin/com/playground/auth/security/SecurityIntegrationTest.kt
@@ -102,7 +102,7 @@ class SecurityIntegrationTest(
             ProductSaveRequestDto(
                 name = "새 상품",
                 stock = 10,
-                price = 10000L,
+                price = 10000.toBigDecimal(),
             )
 
         mockMvc
