+++ module-app/src/test/kotlin/com/playground/product/controller/ProductUpdateApiTest.kt
@@ -23,8 +23,8 @@ class ProductUpdateApiTest(
     @Test
     @WithMockUser(roles = ["ADMIN"])
     fun `상품 수정 - 성공`() {
-        val savedProductJpaEntity = productRepository.save(ProductJpaEntity(name = "상품1", stock = 10, price = 10000L))
-        val updateRequest = ProductUpdateRequestDto(name = "상품2", stock = 20, price = 20000L)
+        val savedProductJpaEntity = productRepository.save(ProductJpaEntity(name = "상품1", stock = 10, price = 10000.toBigDecimal()))
+        val updateRequest = ProductUpdateRequestDto(name = "상품2", stock = 20, price = 20000.toBigDecimal())
 
         performAndDocument("상품 수정 - 성공") {
             httpMethod = HttpMethod.PATCH
@@ -58,8 +58,8 @@ class ProductUpdateApiTest(
     @Test
     @WithMockUser(roles = ["USER"])
     fun `상품 수정 - 실패, USER 권한으로 ADMIN API 접근 시도`() {
-        val savedProductJpaEntity = productRepository.save(ProductJpaEntity(name = "상품1", stock = 10, price = 10000L))
-        val updateRequest = ProductUpdateRequestDto(name = "상품2", stock = 20, price = 20000L)
+        val savedProductJpaEntity = productRepository.save(ProductJpaEntity(name = "상품1", stock = 10, price = 10000.toBigDecimal()))
+        val updateRequest = ProductUpdateRequestDto(name = "상품2", stock = 20, price = 20000.toBigDecimal())
 
         performAndDocument("상품 수정 - 실패, USER 권한으로 ADMIN API 접근 시도") {
             httpMethod = HttpMethod.PATCH
