+++ module-order-contract/src/main/kotlin/com/playground/order/contract/domain/vo/ProductInfo.kt
@@ -1,7 +1,9 @@
 package com.playground.order.contract.domain.vo
 
+import java.math.BigDecimal
+
 data class ProductInfo(
     val productId: Long,
-    val price: Long,
+    val price: BigDecimal,
     val productName: String,
 )
