+++ module-order/src/main/kotlin/com/playground/order/domain/model/OrderProduct.kt
@@ -1,8 +1,10 @@
 package com.playground.order.domain.model
 
+import java.math.BigDecimal
+
 class OrderProduct(
     val id: Long? = null,
     val productId: Long,
     val quantity: Int,
-    val price: Long,
+    val price: BigDecimal,
 )
