+++ module-product/src/main/kotlin/com/playground/product/application/port/inbound/command/SaveProductCommand.kt
@@ -1,7 +1,9 @@
 package com.playground.product.application.port.inbound.command
 
+import java.math.BigDecimal
+
 data class SaveProductCommand(
     val name: String,
     val stock: Int,
-    val price: Long,
+    val price: BigDecimal,
 )
