+++ module-product/src/main/kotlin/com/playground/product/application/port/inbound/command/UpdateProductCommand.kt
@@ -1,8 +1,10 @@
 package com.playground.product.application.port.inbound.command
 
+import java.math.BigDecimal
+
 data class UpdateProductCommand(
     val id: Long,
     val name: String,
     val stock: Int,
-    val price: Long,
+    val price: BigDecimal,
 )
