+++ module-product/src/main/kotlin/com/playground/product/domain/event/ProductCreatedEvent.kt
@@ -1,10 +1,11 @@
 package com.playground.product.domain.event
 
 import com.playground.common.event.DomainEvent
+import java.math.BigDecimal
 
 data class ProductCreatedEvent(
     val productId: Long,
     val name: String,
     val stock: Int,
-    val price: Long,
+    val price: BigDecimal,
 ) : DomainEvent
