+++ module-product/src/main/kotlin/com/playground/product/domain/event/ProductUpdatedEvent.kt
@@ -1,6 +1,7 @@
 package com.playground.product.domain.event
 
 import com.playground.common.event.DomainEvent
+import java.math.BigDecimal
 
 data class ProductUpdatedEvent(
     val productId: Long,
@@ -8,6 +9,6 @@ data class ProductUpdatedEvent(
     val newName: String,
     val oldStock: Int,
     val newStock: Int,
-    val oldPrice: Long,
-    val newPrice: Long,
+    val oldPrice: BigDecimal,
+    val newPrice: BigDecimal,
 ) : DomainEvent
