+++ module-product/src/main/kotlin/com/playground/product/presentation/request/ProductUpdateRequestDto.kt
@@ -2,6 +2,7 @@ package com.playground.product.presentation.request
 
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.PositiveOrZero
+import java.math.BigDecimal
 
 data class ProductUpdateRequestDto(
     @field:NotBlank(message = "{product.name.not-blank}")
@@ -9,5 +10,5 @@ data class ProductUpdateRequestDto(
     @field:PositiveOrZero(message = "{product.stock.positive-or-zero}")
     val stock: Int,
     @field:PositiveOrZero(message = "{product.price.positive-or-zero}")
-    val price: Long,
+    val price: BigDecimal,
 )
